refactor(workersvc): move message ingest + usage fold to usage_fold.go
Pure same-package code motion (PRD #921 M2, seam 4). Moves AppendMessages,
NoteOversizeBatch, appendMessages, foldRunUsage, numericUSD, nonNegTokens and
the appendObservation type out of service.go into the new usage_fold.go,
byte-identically. The shared truncateRunes func and the rune-limit const block
stay in service.go. The now-unused math, math/big and pgx/v5/pgconn imports
are dropped from service.go (pure deletions; their sole users moved). No
exported-API change, no behavior change; go doc output unchanged.
